# [runtime] move simd int->string functions to cpp file