… common/algorithms/simd-int-to-string.h