Surprisingly, these big functions were inlinable and C++ compiler
inlined many calls to int->string conversion functions.

This results in a major code bloat, since int->string conversions
happen in a lot of places in PHP (explicitely or implicitely).

The most troublesome function is `uint32_less10000_to_string`,
it was among the top symbols from the binary size profile.

Moving these functions from a header file to a cpp file reduces
the binary size by several megabytes (!). Depending on the code
base size, the effect can reach ~13000 kb.

Performance-wise, since inlined code is compiled as a part of the
KPHP application code, we don't get performance penalties for
doing an extra function call. Microbenchmarks show that we
can even get some minor speedup in some cases.

	name                       old time/op  new time/op  delta
	Strconv::StrvalInt32small  20.5ns ± 2%  20.0ns ± 0%   -2.44%  (p=0.033 n=10+10)
	Strconv::StrvalInt32       52.4ns ± 1%  52.7ns ± 1%        ~  (p=0.484 n=10+9)
	Strconv::StrvalInt64       72.1ns ± 3%  58.0ns ± 0%  -19.56%  (p=0.000 n=10+9)
	[Geo mean]                 42.6ns       39.4ns        -7.60%